Create a new component for the switch theme handler
