+++ src/screens/HomeScreen.tsx
@@ -1,11 +1,9 @@
-import React, { useContext, useState } from 'react';
+import React, { useContext } from 'react';
 import {
   ActivityIndicator,
   FlatList,
   Image,
-  Platform,
   StyleSheet,
-  Switch,
   Text,
   View,
 } from 'react-native';
@@ -13,26 +11,17 @@ import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import Icon from 'react-native-vector-icons/Ionicons';
 
 import { GistCard } from '~src/components/GistCard';
+import { SwitchTheme } from '~src/components/SwitchTheme';
 import { ThemeContext } from '~src/context/theme/theme';
 import { useGists } from '~src/hooks/useGists';
 import { globalStyles } from '~src/styles';
 
 export const HomeScreen = () => {
-  const { theme, setLightTheme, setDarkTheme } = useContext(ThemeContext);
+  const { theme } = useContext(ThemeContext);
   const { top } = useSafeAreaInsets();
   const { isLoading, publicGists } = useGists();
-  const [darkIsEnabled, setDarkIsEnabled] = useState<boolean>(theme.dark);
 
   const { colors } = theme;
-
-  const toogleSwitch = () => {
-    setDarkIsEnabled(!darkIsEnabled);
-    if (theme.dark) {
-      setLightTheme();
-    } else {
-      setDarkTheme();
-    }
-  };
 
   return (
     <View
@@ -45,24 +34,7 @@ export const HomeScreen = () => {
           Public Gists
         </Text>
         <Icon size={30} color={colors.text} name="git-branch-outline" />
-        <View style={styles.themeSwitchWrapper}>
-          <Icon name="sunny-outline" size={25} color={theme.colors.text} />
-          <Switch
-            style={styles.switch}
-            value={darkIsEnabled}
-            onChange={toogleSwitch}
-            trackColor={{
-              false: theme.colors.text,
-              true: theme.colors.text,
-            }}
-            thumbColor={
-              Platform.OS === 'android'
-                ? theme.colors.border
-                : theme.colors.background
-            }
-          />
-          <Icon name="moon-outline" size={25} color={theme.colors.text} />
-        </View>
+        <SwitchTheme />
       </View>
       <Image
         source={require('~src/assets/images/octocat.png')}
@@ -98,12 +70,7 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     marginBottom: 10,
   },
-  themeSwitchWrapper: {
-    alignItems: 'center',
-    flex: 1,
-    flexDirection: 'row',
-    justifyContent: 'flex-end',
-  },
+
   textTitle: {
     marginRight: 10,
     fontSize: 35,
@@ -118,8 +85,5 @@ const styles = StyleSheet.create({
   },
   gistsContainer: {
     flex: 1,
-  },
-  switch: {
-    marginHorizontal: 2,
   },
 });
